提交今天的进度

+++ phpstr.h
@@ -5,5 +5,5 @@
 #ifndef CURRICULUM_DESIGN_PHPSTR_H
 #define CURRICULUM_DESIGN_PHPSTR_H
 int strpos(char *haystcak,char needle, int offset);
-char * substr(char *string, int start, size_t length );
+char * substr(char *string, int start, int length );
 #endif //CURRICULUM_DESIGN_PHPSTR_H
